bug here

+++ include/Config.hpp
@@ -12,7 +12,7 @@ namespace lq{
     template<typename T>
     struct is_printable<T, std::void_t<decltype(std::declval<std::ostream&>() << std::declval<T>())>> : std::true_type {};
 
-    using label = int ;
+    using label = short ;
     // Static assert to ensure label is printable
     static_assert(is_printable<label>::value, "label type must be printable");
 }
